feat: 1450 ThorClient refactoring
