+++ packages/network/src/thor-client/contracts/contracts-module.ts
@@ -12,7 +12,7 @@ import { type Abi } from 'abitype';
 import { type VeChainSigner } from '../../signer/signers/types';
 import { BUILT_IN_CONTRACTS } from '../../utils';
 import { decodeRevertReason } from '../gas/helpers/decode-evm-error';
-import { type ThorClient } from '../thor-client';
+import { type ThorClient } from '../ThorClient';
 import {
     type SendTransactionResult,
     type SimulateTransactionOptions
