+++ packages/network/src/thor-client/nodes/nodes-module.ts
@@ -1,7 +1,7 @@
 import { NODE_HEALTHCHECK_TOLERANCE_IN_SECONDS, thorest } from '../../utils';
 import { InvalidDataType } from '@vechain/sdk-errors';
 import { type CompressedBlockDetail } from '../blocks';
-import { type ThorClient } from '../thor-client';
+import { type ThorClient } from '../ThorClient';
 import { type ConnectedPeer } from './types';
 import { HttpMethod } from '../../http';
